feat: GitHubReleaseFetcher
